+++ libs/ui/src/app/core/constants/app-version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '23.0.1';
+export const APP_VERSION = '23.0.2';