+++ libs/ui/src/app/shared/components/fieldset/fieldset.ts
@@ -1,7 +1,12 @@
-import { Component } from '@angular/core';
+import { Component, input } from '@angular/core';
+import { FieldState } from '@angular/forms/signals';
+import { ValidatorHint } from '../validator/validator-hint';
 
 @Component({
   selector: 'app-fieldset',
   templateUrl: './fieldset.html',
+  imports: [ValidatorHint],
 })
-export class Fieldset {}
+export class Fieldset {
+  field = input.required<FieldState<any, any>>();
+}